fix(classe + cheat): corrigida escolha de classe e Fake Godmode
== Characters

- Corrigido bug da classe não estar a ser inicializada.

== Cheat

- Corrigido godmode com valores inválidos
(Problema com int e long long int).

+++ src/entity/entity.h
@@ -15,8 +15,8 @@
 
 typedef struct entity {
     Coords coords;
-    int maxHealth;
-    int health;
+    unsigned int maxHealth;
+    unsigned long long health;
     int armor;
     int basedamage;
 } ENTITY, *Entity;
